pyproject.toml
[project]
name = "python-5-note-making-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.0"
dependencies = [
    "faker>=33.0.0",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.7",
    "sqlmodel>=0.0.22",
]
